+++ CHANGELOG.md
@@ -12,6 +12,7 @@ contents to a new release section (e.g. _[0.x]_).
 ### Added
 
 - Normalize equations with logarithmic terms. (PLUT-1120)
+- Solve elementary logarithmic equations. (PLUT-1123)
 
 ### Fixed